# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Portfolio website API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List all skills
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{id}:
    patch:
      operationId: updateSkill
      tags: [skills]
      summary: Update a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillUpdate"
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === PROJECTS ===
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === EXPERIENCE ===
...
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List all testimonials
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create a testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    patch:
      operationId: updateTestimonial
      tags: [testimonials]
      summary: Update a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialUpdate"
      responses:
        "200":
          description: Updated testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === BLOG ===
...
components:
  schemas:
...
    Skill:
      type: object
      required: [id, name, category, level, icon, sortOrder]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        level: { type: integer }
        icon: { type: ["string", "null"] }
        sortOrder: { type: integer }
        createdAt: { type: string }
    SkillInput:
      type: object
      required: [name, category, level]
      properties:
        name: { type: string }
        category: { type: string }
        level: { type: integer }
        icon: { type: string }
        sortOrder: { type: integer }
    SkillUpdate:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        level: { type: integer }
        icon: { type: string }
        sortOrder: { type: integer }

    # === PROJECT ===
...
    Service:
      type: object
      required: [id, icon, title, description, sortOrder, createdAt]
      properties:
        id: { type: integer }
        icon: { type: string }
        title: { type: string }
        description: { type: string }
        pricing: { type: ["string", "null"] }
        features: { type: array, items: { type: string } }
        sortOrder: { type: integer }
        createdAt: { type: string }
    ServiceInput:
      type: object
      required: [icon, title, description]
      properties:
        icon: { type: string }
        title: { type: string }
        description: { type: string }
        pricing: { type: string }
        features: { type: array, items: { type: string } }
        sortOrder: { type: integer }
    ServiceUpdate:
      type: object
      properties:
        icon: { type: string }
        title: { type: string }
        description: { type: string }
        pricing: { type: string }
        features: { type: array, items: { type: string } }
        sortOrder: { type: integer }

    # === EXPERIENCE ===
...
    Testimonial:
      type: object
      required: [id, name, role, company, quote, sortOrder]
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        company: { type: string }
        quote: { type: string }
        avatarUrl: { type: ["string", "null"] }
        rating: { type: integer }
        sortOrder: { type: integer }
        createdAt: { type: string }
    TestimonialInput:
      type: object
      required: [name, role, company, quote]
      properties:
        name: { type: string }
        role: { type: string }
        company: { type: string }
        quote: { type: string }
        avatarUrl: { type: string }
        rating: { type: integer }
        sortOrder: { type: integer }
    TestimonialUpdate:
      type: object
      properties:
        name: { type: string }
        role: { type: string }
        company: { type: string }
        quote: { type: string }
        avatarUrl: { type: string }
        rating: { type: integer }
        sortOrder: { type: integer }

    # === BLOG POST ===
...