# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio website API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  icon: string;
  title: string;
  description: string;
  /** @nullable */
  pricing?: string | null;
  features?: string[];
  sortOrder: number;
  createdAt: string;
}