# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio website API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: number;
  name: string;
  category: string;
  level: number;
  /** @nullable */
  icon: string | null;
  sortOrder: number;
  createdAt?: string;
}